ProjectManagerPlugin: beginning of actual project management
There is a defintion of a json file, which tells the IDE
how to inteprete this dir. How to build (or other targets
you define), which are the binaries you want to execute, etc.

The IDE will load that file, and eventually know how to handle
it to build the project, and run your executable.

For now - you can just see the files in each project's directory.

If the project contains a "qtedit4.json" it is parsed, and the fist
command/executable are marked. Still nothing is beeing run.