All: Fix displayig of icons on Windows/Linux
Windows: the external project support just did not work. I was unsure
why - do I re-did that function in good old simple
download/extract/copy. The root of the problem was (?) that the icon
theme contains logs of symbolink links, which messed up the installer.
To overcome this, I also add some paths in the installer, not the whole
icon set. I would like to revisit this in the future.

Linux: It seems that the app image creation tool I am using (there is
another, but it has its own problems) does not bundle the icon plugin.
So displaying icons on the app image would fial. Instead - I copy them
manually (in the github action and build script), as well as the wayland
plugins.

This took 56 iterations over 2 weeks of hard detective and undocumented
work. FUCK THIS.