Windows: able to compile the project
I added ways to add paths, and call a setup script (this does not work).
Just adding path worked for me and I was able to properly execute a
build.

Running still failed, as qtedit needs QT which is not in path.

Which means I am going to scrape this, and start working on Kits from
now on.