KitDetector: fix compilation on windows
1) `getenv()` should not be used on windows. This copies the global
environment variable to the process, and does not get updated. Instead
`GetEnvironment()` should be used. Wrapped this in a function that
returns `std::string` to create a copy (windows needs a buffer, and Unix
returns a pre-allocated buffer).
2) I was missing `sstream` include.
3) When iterarting over the path, the CL compiler behaves differently.
You cannot append directly a string to a path. OK. Seems resonable.
4) When working on windows - use the proper path separator, and add the
exe suffix.
5) Removed extra indent in the loop, because WTF not.