CMake/Meta: work offline
Assume you are on a trip, and want to work on the code of the project.
Before leaving, run this script, and the configure with
`QTEDIT4_WORK_OFFLINE=on`.

Now you should be able to work offline (the CMake stuff will not query
online servers. When you are done - for each of the repos you modified,
create a new branch, and push and create PRs.