18 - keep original line endings
Ideally - the line endings would have been saved inside each paragraph.
However, Qt in its wisdom desided to nuke the line endings on each
QTextBlock.

I tried adding a block data with the origianl line endings when loading
the text, however:

1) Using a cursor to add lines manually is *too* slow. Loading file with
1000 lines takes a few minutes (!)
2) The internal architecture of Qutepart refuses it, as the internal
state of the data must be valid.

Instead, I opted to detect the first line, and assume the rest will
follow. This again was tricky, as the text stream used to open the files
was "TEXT" and this caused Qt (again in its holy wisdom) to eat up "\r"
characters. Puting the file in binary breaks nothing, and it fixes this
problem.

Also fixed the problem of the notification about the file beeing changed
while saving.

Also properly inlined some functions.

Also - made sure that the "Plain text" syntax is the first of the list.