TextEditor: search history - 39
The find/replace widets for the text editor are moved to the bottom of
the editor (also, the banner is moved on top of the editor). Both were
previously inline, which sounded cool, but just did not work.

Created a new line text edit, which has a history option. I also made
the history a shared data - which means that if you seach for something
in one window, it will present in the history of another editor.

The search history is saved as part of the config file.

closes #39