OutputDetector: process line only once
Each output line can be parsed/decoded/detected using a single detector.
Once the GCC output detector has worked, the golang does not need to
work.

Before, this broke on some projects, as an output could be part of a
golang error, and then the GCC output (which contains full path) would
be treaten as a relative path - would be effectively contain the project
path twise, and thuse messing the issues display.

Fixes #43