qmdiEditor, HexViewrPlugin: notify loading of a text file
In preparation for #52, lets announce to the plugins "hey - a new file
has been loaded". When the project manager will decide "if this project
has not been loaded, lets ask the user".

The hex viewer plugin was eating those requests, so it needed fixing.

This is a new way of using the plugin system, which has not been tested
yet, but it was the intention when developing it.