ProjectIssuesWidget/ProjectManagerPlugin: scroll to error/warning output (#48)
When you click an issue in the issues widget, not only the editor will
be open in the correct place, but the relevant build output will be
visible (the output panel will scroll in such way that the corresponding
line of the build will be visible).

I whish I could put that line at the top, but Qt's API is just lame.

closes #48