ProjectManagerPlugin: fix double entries of files when loading (#58)
When loading a project, the files were added manually to the list. This
was wrong, as the list of files should be modified only when the combo
changes.

Added also a loading indicator - which is not working. I think the
design of this part is currently broken. Wil need to fix this
eventually.

Closes #58