main: split actions visible
We are modifyinf built in actions of the main window. Unless we properly
reset this, the actions will not be available in minimized mode.

Now split action does not need to be manually added to the main window.
This could have been done with another plugin (I might move the split
tab to a seperate plugin, in case it miss behaves - this will also
work).