SplitTabWidget: Starting to work on split tabs (#54)
New widget+feature: split tabs.

You can press "control+\" to create a new split tab, with an empty text
editor. if you press "control+shift+\" the current tab is moved to a new
split.

This is specially usefull in C/C++ with headers. You pres F4 to get into
the header, then press "control+shift+\" and the header is sent  to a
left split.

Issues still missing:
https://github.com/diegoiast/qtedit4/issues/59
https://github.com/diegoiast/qtedit4/issues/60
https://github.com/diegoiast/qtedit4/issues/61
https://github.com/diegoiast/qtedit4/issues/62
https://github.com/diegoiast/qtedit4/issues/63

and probably more. I want this issue in the main code now, so not
perfect - but working good enough to get in.

closes #54