QPathEdit: default to inline button (#68)
The original code defaults to seperated button. In our implementation,
the first thing we do, is set it to inline. For some reason - the
upstream version messed up GNome, and the maximize funciton of windows
is disabled.

On KDE - it still pops up a window randmly. I have no idea why.

This code is not good enough, and will need to be removed ASAP.

Fixes #68