qmdiEditor: display tooltip for current symbol
When you hover a word, the editor will query the plugin system for meta
data about this symbol.

Currently - the ctags plugin returns what it knows about it. Type is
missing. Location is in vi/ed command - which means "goto source" will
have to search in the source for for that.

The ctags plugin will create a tags file on each build. It will also
loade that when the project is loaded.

Problem: we need to have a list of tag files, and not a single one. When
an editor asks for metadata - we need to look for the most suited for
it, and then query.

We need a new singal, "project removed" to clear memory.

We need also to format the editor tooltip a little more pretty.