CTagsLoader: fix loading of extended info (#66)
UCTags extended info is working now. It seems no LLM can get it right,
so I did it the old fasion way - and coded it manually. Now I even
understand how it works and the original model was completlly broken.

This finally closes issue #66 - and we can display tooltips about a
symbol in our code.

closes #66