SplitTabWidget: new tab should be opened on correct split (#77)
When you click the new button on the corresponding split, a new text
editor should open on that split - not the active.

This was not that trivial, but now it works.

I added support for close buttons on tabs - but I decided against this,
as it will take too much visible space. Users can right click the tab,
or right click and choose the corresponding command.

closes #77