qmdiTabWidget: App button
In minimal mode, the mdi server will display a different main button.
Lets call it the AppMenu. Its like a humburger, but it contains the
full menu, with nested sections. Its built from the same actions list
as the full menu.

Pressing alt+m when the menu is open, will close it.

I also have a code for a more stylized button, but not using it ATM,
as I am unhappy about its look. Fow now, its a simple QPushMenu with
a tint. And that works good enough for now.