ProjectManagerPlugin: after build, cpu is at 100%
After a build, on linux, there was a CPU core stucked at 100%. This is a
regression caused by 348d1acc687439883e1a015fbae63041524d0f3a.

Problem is that the socket used to read from the process continued to
read, even after the process died. This is a bug caused by Perplecity's
code. Vibe coding FTW.