CTags: speed improvements
1) When quering - move the if to the top, instead of inside the loop.
Now that if is done only once and not 100,000 times

2) The ctags entry : maintain the original string, and then all fields
are string views into that original line. This means less memory is
used, and less copies.

The last step reduced loading of:

```
-rw-rw-r-- 1 diego diego 563M Apr 25 11:36 build/qtedit4.tags

```

From 6 seconds, to 2.5. Memory consumption is about 1GB. Meaninng: I am
creating too much tags, and laoding needs to be done in another thread.

This is still a win.