ProjectManager: fix for building on multiple Qt6 kits
For some reason, I was trying to build using Qt 6.9.0 on MSVC, and on my
path, before the 6.9.0 I see Mingw 6.8.3. This broke building *inside*
qtedit4, until I found the solution/problem.

CMake will use Qt from the first qmake it finds in path. But The Qt I
was using, is not the first. All the environment variables I was using,
were not the ones CMake was looking for.

Solution: add another environment variable. Now it works :)