SplitTabWidget: fix some crashes
This patch fixes 2 problems:

- When saving an empty file, the code crashed.
- When opening a save dialog, over a split with a single tab - that said
tab/split is closed, when you cancel the saving.
- (found later on): when you close the last tab, and you have a single
split, when creating a new tab the app would crash. This is because the
last active tab was not reseted, and it became a dangling pointer.

Seems to work on MSVC, and MingGW. Will test on Linux "soon".