CTagsLoader: speed up query (#86)
Previous code did a linear scan over the tags database. On large
databases (1967979 items in my case), the query took 2 seconds.

As the tags file is already ordered, there is no need for linear scan.
We can do a binary search (or somthing similar), to speed things up.

This fix reduces searches from 2.6 seconds (!) to 0ms (!) while
searching in this DB for a 3 letter word.

No further refactors are needed *at this time*.

closes #86