SplitTabWidget: fix focus on split panels
Split the panel in two, open file1, and file2 (on the second panel).

If you select now "file1" (the left panel) and open a new file - the new
file was added to the right panel. This is because we were not able to
detect that the left QTabWidget is active.

This commit fixes that issue, see also https://github.com/diegoiast/
qtedit4/issues/61 which makes it hard to see which editor is active.