SplitTabWidget: simplify closing on empty split
Previously, I was listining to events and guess when a child was
removed, in hope that I get the last tab removed from a tab widget.

Instead - I can just use `QTabWidget::tabRemoved(int)` to get notified
on this event. This removes a lot of ugly code, and fixes some crashes.

Also - fixes the dragging of last tab from a split. Before, if a single
tab was available, that split would be closed. Now its not (since we can
explicitly ask "do we have 0 tabs left?", before we neede to ask for
"1").