ProjectManagerPlugin: fix crash when no kits are found
I found this bug in a live session of Debian. I used the appimage, and
did not understand why the app crashed.

I manged to trace it down in a Win10 installation, that uses mingw
(which is a pain to debug, as it did not catch the exception, and
debugging loading takes about a minute).

Anyway, when no kits are available, at startup the combo box of the kits
will be set to 0, and will de-reference. Its actually an out of boundery
- we try to get element "0" when we have no elements. Just did an if,
and code no longer crashes.

Doing a rescan should be done as a first step - see https://github.com/
diegoiast/qtedit4/issues/25