Github actions: install specific clang version
It seems that the image that github provides has clang18, and a newer
visual studio. Both of this collide. From now on, a specific clang is
installed.

This is reported here: https://github.com/actions/runner-images/issues/12435

This is done only on Windows, as Linux/OS have no such problems.