qmdiSplitTab: properly support "new"
When clicking ghenew button from the menu, previously 2 new files were
added. The reason was that I hacked inside the split tab to accept the
new file command.

The correct solution is to leave it to the default `addTab()` method.
However, this broke another thing (which I tried to avoid in the
previous hack): if you click new in a split which was not active, I was
expecting that split to add a new file. Instead - a new tab was added
inside the active tab.

Solution: inside the click handler of the tool buttons, make that tab
active. Now everything works as expected. This is the code in
qmdiSplitTab.

Code in SplitTabsPlugin was removed, as it was wrong to begin  with.

All other code is random cleanups.