KitDetector: fix tasks creation
Since we moved from a single command to a list of commands, `eval` is
the way. Before this fix, and the multiple command - linux did not work
at all.

Also, we want to for for this environment to use the new Qt libraries -
do LD_LIBRARY_PATH was added. And also add support for macOS. Not
tested, but should work.