ProjectManagerPlugin: run applications from kit
We gained the ability to run the executable itself, as it was a task.
This gives the option to copmile using the system qt, and then run it
against several other Qt installed.

Work on linux so far. Windows testing will follow soon.

By now, its disabled to simplify things, as its not tested enough.