ProjectManager: speed up log output
In previous code, I converted output to HTML, and then added HTML to the log output. This is extreamly slow.

Instead, I now use the Qt API to create the colored text.

Before this fix, while building qtedit4 on windows, on itself, the UI would freeze for 5-10 minutes. Now it freezes for less then 1 minute. I know it can be reduced a little more, if I just add plain text, and then assign a syntax highlighter.

lets see how this works for now, before I choosse a new path, as this code is still not usable (cleanups are a must, relative paths needs to be fixed).