PluginManager: port UI to use a new file list widget (103)
Made a new widget which deals with loading a directory asynchronously.   This reduced a lot of complexity in the project manager plugin (as it contained some part of this logic spread in the actual code).

The new files list scans dir in a thread, and is fast/efficient to be used while debugging in MSVC in debug mode. Before, this part of code was practically a pain, as it was slow.

I removed an old work-around for a bug I had in a previous version where selecting a new project was called several times.

Fixed #103 however, the command palette is still not working.

`LoadingWidget.h` got renamed to `hpp,`, as its requested by newer C++ code.