ProjectBuildConfig: fix configuring a configured project (#108)
When we re-configure an existing dir, on windows it would fail. The
reason is `mkdir  ...` fails, as the directory already exist.

On Unix we can use `mkdir -p` to overcome it. On Windows, the solution
was to pipe the command trough another shell. Maybe in the future I will
revisit it in power shell, in hope that this will work nicer.

How did I debug this? I added "echo" before each command to find out
which one failed.

```
            {"echo 1111111111111111111111111",
             "cmd /c \"mkdir \"${build_directory}\\.cmake\\api\\v1\
\query\" >nul 2>nul || rem\"",
             "echo 2222222222222222222222222222",
             "cmd /c \"type nul > \"${build_directory}\\.cmake\\api\\v1\
\query\\codemodel-v2\" || "
             "rem\"",
             "echo 3333333333",
             "cmake -S \"${source_directory}\" -B \"${build_directory}\"
"
             "-DCMAKE_BUILD_TYPE=Debug"});
```

closes #108