CTagsPlugin: load tags in a background thread (109)
Loading the ctags files should not be done in the main thread. We now
load it in a background thread. This reduced on my setup 1.5 seconds on
start up.

In theory, this is bad, since others might consume this API (the ctags
completion). But, in reality this is a read operation... and in such
case users will get "non ideal results", then they will re-try and it
will work.

closes #109