KitDetector: fix finding cargo in system path
In some setups (Arch linux, when you `pacman -S cargo`) there will be no
CARGO environment, but the binary will just be in the $PATH.

Lets support this scenario, by looking for the binary in the path when
`CARGO_HOME` is not avialble. Should also work on Windows, which is
nice.