ProjectManagerPlugin: fix loading of files in the output (#116)
When an error is spit by the build system of the language, the emmited
filename is absolute (MSBuild), or relative to the source dir (CMake,
cargo).

And then there is golang. Which uses `./main.go` in the error message.
Which the code identified as "absolute". So, from now on, we use source
dir as the base. And detect `./` as relative.

refs #116. I am not closing it, since the code which handles links
should always deal with absolute paths. The linkify path should set the
links to be absolute, while decoding the output (not at click time).