Meson: try to run meson command using kit
While detecting the project files targets, we run cargo and ask it to
output in json format the project structure. This does not work, as
cargo might be in the kit's path and not system path.

This fix tries to fix it by running the cargo command using the kit.
So far, this is not working as I expected.