ProjectManagerPlugin: file manager - add dir as project command menu (#11)
This adds support for a new command visible on the file sub menu. It is
done by providing an extra command using `getContextMenuActions()`.

Used this to refactor the loading code, we can now just iterate normally
instead of the mostrocity we had before which used timers. The code is
fast now anyway.

closes #11