Cargo: find available packages
Instead of running the cargo command manually, lets create the metadata
json right inside the kit. This is similar to what is been done in
cmake.

In cmake, we have an API that create files. In cargo - we just redirect
the output to a file. Then - instead of reading stdout, we read the
file.

Simple - and works. On linux. I might need to fix it on Windows.