ProjectManagerPlugin: load project when a file is loaded (122)
When a file is loaded into the editor, if the file represents a project
we support (go.mod, .build.meson, cargo.toml, cmakelists.txt,
qtedit4.json and probably more in the future) - we ask the user if we
wants to load also the project, or just load the file for editing.

We just extend a previous functionality we had, which worked only for
our project files.

closes #122