ProjectManagerPlugin: add run action to mdi clients
When an editor gets loaded, it announces about the new filename. Then,
the project manager plugin will inject into it a new action.

This surprisingly works. But has a huge problem: we don't know which
client triggered it, as its a global object. (right click on a non
active tab - and the current tab gets triggered anyway).

We need to add to each client a new action, and that will be also added
to the project menu. This way, we know the called client (instead of
using `data()` to store state).