PluginManager: add run action to each qmdiEditor
Moved the global action to be local to each editor. This simplifies code
a lot. Now, the new command is also visible at the project menu, and can
be executed using a key shortcut (still does nothing, but now its
trivial to fix).

Needs a new qmdilib version due to new API to query action groups.