FilesList: use a simpler list
Instead of using a QListItem, this code will use a new model, which is
an enhanced QStringModel. This reduced memory footprint in my test case
(a dir with 130k files). In this measurement, I open the ide with no
project, load that project, unload and load again. I measured the memory
usage as displayed by `btop` (in MB).

Before:
 * before load     176M
 * after load      468M
 * after unload    442M
 * after load2     581M
 * after unload2   551M

After:
 * before load     177M
 * after load      379M
 * after unload    338M
 * after load      469M
 * after unload    367M

I see that after the unloads the memory does not drop back to original
values. I am unsure if this is a bug (memory leak) in my code, or some
internal Qt cache (as values differs between runs).

This is a good fix anyway.