Meta: build flavors
We have 3 different builds. From IDE will be yellow, named "dev".
Then for release we have 2, blue for community edition, and green for
official.

`build.sh` is known to work with those 2 new builds. Windows and github
are not verified yet.

Distributions should use the "ce" build. Which should integrated with the
environment (less shading, and gradients).