qmdiEditor/Unix: use excutable bit for scripts (#137)
We previously used the filename (on unix!) for recognizing if the file
can be executed. This was very wrong. Lets see if the "x" bit is set
for the file.

When saving "as", this means a new filename, we will automatically
make the file "executable", if it starts with a shebang. This fixes
most problems.

We might do this on every save.
We might do this per configuration.

For now, if you save a file first, and it contains no shebang
(for example, python code copied from chatgpt), it will not be
made executable. Added the shebang will do nothing (until you "save as"
even under the same name).

closes #137