qmdiSplit: do not open a new file on every start
Problem: we modify the minified mode, which in turn will trigger the
split tab to create a new file. However, on this stage we did not
finished loading.

Simple Qt trick to disable the signal. Alternative is to make the split
not to accept new tabs, and expose it through the plugin.

This is simpler.