Flatpack: initial support (#24)
Added the needed parts to at least build locally the flatpack.
Adding the `appdata.xml` was very beneficial, as it added more
metadata to the AppImage.

Offline support got fixed, as sub projects gained offline support, and
also - we are no longer installing the icons on such scenarios.

Not a full win, as I there is no automatic way to sync the subprojects
in the cmake to the build yml, there is no support for build flavors,
and the build is no upstream.

But, its a good start.

refs #24